requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask==3.1.1
openpyxl==3.1.5
python-dotenv==1.0.1
requests==2.31.0
